refactor!: use `fmt::Write` for `XmlWriter`
This refactoring primarily does three things:
1. Removes the mandatory lifetime from `XmlWriter`;
2. Significantly reduces the number of allocations in serialization functions;
3. Opens the door to `#![no_std]`. I don't know what kind of scenario could possibly benefit from that, but if we don't *need* `std`, why not just make it `#![no_std]`?

Since we are only ever dealing with UTF-8 in and UTF-8 out, `io::Write` never really made sense. While this would be useful for writing directly to some kind of networking payload, for instance, many uses of SSML would require serializing to a string to put in some other `serde` structure, or in a text payload (as is the case with acss-rs) anyway.